requirements.txt
pymongo
web3
python-dotenv
requests
tenacity 